Add opt-in richer HSTS support;
This patch adds [opt-in] richer HSTS support. We've had to add these properties
for our Nginx configuration a few times in the past. It would be nice to
have the buildpack include the whole header rather in our custom location.conf so
we can rely on a more feature-rich version of what shipped in [v1.3.13](https://github.com/cloudfoundry/staticfile-buildpack/releases/tag/v1.3.13).